+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### Added
+
+- Support for UPduino 3.1
+  - Image based on UPduino assembly (https://github.com/tinyvision-ai-inc/UPduino-v3.0/blob/master/Board/v3.0/UPduino_v3.0Assembly.html ; Open locally to render)
+    - Selected FPGA, export Front with default (1000x1000); Use gimp to crop to png
+    - Save to src/webview-js/board-layout-upduino31.png
+
 ## [0.0.19]
 
 ### Added